Remove Deprecated Microsoft.Extensions.PlatformAbstractions
